refactor(simulator-network): typealias NodeId is now a Long + playground refactor

+++ opendc-simulator/opendc-simulator-network/src/main/kotlin/org/opendc/simulator/network/components/Node.kt
@@ -14,7 +14,7 @@ import org.opendc.simulator.network.utils.logger
 /**
  * Type alias for improved understandability.
  */
-internal typealias NodeId = Int
+internal typealias NodeId = Long
 
 /**
  * Type alias representing the routing vector of a [Node]. For improved readability.
